### 2025-05-17 #8237 @arfon
@ohno – thanks for your submission to JOSS. From a quick review of the paper, it seems like the focus of this software is primarily teaching/understanding. If so, then this submission would not be eligible to be published here.

I'm going to start a scope review amongst the JOSS editors to decide if this submission is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS.

This process will likely take a week or so to complete.
