### 2025-05-17 #8237 @ohno
@arfon Thank you for your quick review. While this software is also useful for teaching and understanding, the focus of this software is primarily software testing in quantum mechanical calculations.

For example, it has already been used to test [TwoBody.jl](https://github.com/ohno/TwoBody.jl), a package for solving two-body problems. Furthermore, the original code underlying this package was used to verify calculations in a peer-reviewed research article on quantum mechanical two-body systems: [A. J. Arifi, L. Happ, S. Ohno, M. Oka, Phys. Rev. D 110, 014020 (2024)](https://doi.org/10.1103/PhysRevD.110.014020). We believe that this illustrates the relevance of the software for research purposes beyond education.

According to [the JOSS definition](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software), this package corresponds to software that offers a mathematical library, or similar. We look forward to a positive outcome from the scope review.
