### 2025-05-27 e9ec13d
Update paper.bib
