+++ joss/paper.bib
@@ -149,88 +149,16 @@
   month = feb
 }
 
-@article{Nagahiro2017,
-  title = {Structure of charmed baryons studied by pionic decays},
-  volume = {95},
-  ISSN = {2470-0029},
-  url = {http://dx.doi.org/10.1103/PhysRevD.95.014023},
-  DOI = {10.1103/physrevd.95.014023},
-  number = {1},
-  journal = {Physical Review D},
-  publisher = {American Physical Society (APS)},
-  author = {Nagahiro,  Hideko and Yasui,  Shigehiro and Hosaka,  Atsushi and Oka,  Makoto and Noumi,  Hiroyuki},
-  year = {2017},
-  month = jan 
+@article{Danisch2021,
+  doi = {10.21105/joss.03349},
+  url = {https://doi.org/10.21105/joss.03349},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {65},
+  pages = {3349},
+  author = {Simon Danisch and Julius Krumbiegel},
+  title = {{Makie.jl}: Flexible high-performance data visualization for {Julia}},
+  journal = {Journal of Open Source Software}
 }
 
-@article{Schrodinger1926,
-  title = {Quantisierung als Eigenwertproblem},
-  volume = {384},
-  ISSN = {1521-3889},
-  url = {http://dx.doi.org/10.1002/andp.19263840404},
-  DOI = {10.1002/andp.19263840404},
-  number = {4},
-  journal = {Annalen der Physik},
-  publisher = {Wiley},
-  author = {Schr\"{o}dinger,  E.},
-  year = {1926},
-  month = jan,
-  pages = {361–376}
-}
-
-@article{Burrau1927,
-  author = {Øyvind Burrau},
-  title = {Berechnung des Energiewertes des Wasserstoffmolekel-Ions (H2+) im Normalzustand.},
-  journal = {Matematisk-fysiske Meddelelser},
-  volume = {17},
-  number = {14},
-  pages = {1-18},
-  year = {1927},
-  URL = {http://publ.royalacademy.dk/books/72/399},
-  eprint = {http://publ.royalacademy.dk/books/72/399}
-}
-
-@article{PrezTorres2019,
-  title = {Dilemma of the “Best Wavefunction”: Comparing Results of the STO-NG Procedure versus the Linear Variational Method},
-  volume = {96},
-  ISSN = {1938-1328},
-  url = {http://dx.doi.org/10.1021/acs.jchemed.8b00959},
-  DOI = {10.1021/acs.jchemed.8b00959},
-  number = {4},
-  journal = {Journal of Chemical Education},
-  publisher = {American Chemical Society (ACS)},
-  author = {Pérez-Torres,  Jhon Fredy},
-  year = {2019},
-  month = mar,
-  pages = {704–707}
-}
-
-@article{Pritchard2019,
-  title = {New Basis Set Exchange: An Open,  Up-to-Date Resource for the Molecular Sciences Community},
-  volume = {59},
-  ISSN = {1549-960X},
-  url = {http://dx.doi.org/10.1021/acs.jcim.9b00725},
-  DOI = {10.1021/acs.jcim.9b00725},
-  number = {11},
-  journal = {Journal of Chemical Information and Modeling},
-  publisher = {American Chemical Society (ACS)},
-  author = {Pritchard,  Benjamin P. and Altarawy,  Doaa and Didier,  Brett and Gibson,  Tara D. and Windus,  Theresa L.},
-  year = {2019},
-  month = oct,
-  pages = {4814–4820}
-}
-
-@article{Hehre1969,
[…]
