+++ joss/inara.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # https://github.com/openjournals/inara
 # Boot "Docker Desktop" before running this script.
-# Run `bash make.sh` on pwsh or WSL.
+# Run `bash inara.sh` on pwsh or WSL.
 docker run --rm -it \
     -v $PWD:/data \
     -u $(id -u):$(id -g) \
