+++ joss/paper.bib
@@ -11,6 +11,19 @@
   url={https://epubs.siam.org/doi/10.1137/141000671}
 }
 
+@article{Danisch2021,
+  doi = {10.21105/joss.03349},
+  url = {https://doi.org/10.21105/joss.03349},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {65},
+  pages = {3349},
+  author = {Simon Danisch and Julius Krumbiegel},
+  title = {{Makie.jl}: Flexible high-performance data visualization for {Julia}},
+  journal = {Journal of Open Source Software}
+}
+
 @article{Gowda2022,
   author = {Gowda, Shashi and Ma, Yingbo and Cheli, Alessandro and Gw\'{o}\'{z}zd\'{z}, Maja and Shah, Viral B. and Edelman, Alan and Rackauckas, Christopher},
   title = {High-Performance Symbolic-Numerics via Multiple Dispatch},
@@ -63,6 +76,15 @@
   author = {Hiyama,  Emiko and Kamimura,  Masayasu},
   year = {2018},
   month = sep 
+}
+
+@book{Cpp,
+  author = {{ISO}/{IEC}},
+  title = {{ISO/IEC 14882:2020 Programming languages — C++}},
+  publisher = {International Organization for Standardization},
+  url = {https://www.iso.org/standard/79358.html},
+  year = 2020,
+  month = feb
 }
 
 @misc{quadgk,
@@ -137,28 +159,3 @@
   adsurl  = {https://rdcu.be/b08Wh},
   doi     = {10.1038/s41592-019-0686-2},
 }
-
-
-
-@book{Cpp,
-  author = {{ISO}/{IEC}},
-  title = {{ISO/IEC 14882:2020 Programming languages — C++}},
-  publisher = {International Organization for Standardization},
-  url = {https://www.iso.org/standard/79358.html},
-  year = 2020,
-  month = feb
-}
-
-@article{Danisch2021,
-  doi = {10.21105/joss.03349},
-  url = {https://doi.org/10.21105/joss.03349},
-  year = {2021},
-  publisher = {The Open Journal},
-  volume = {6},
-  number = {65},
-  pages = {3349},
-  author = {Simon Danisch and Julius Krumbiegel},
-  title = {{Makie.jl}: Flexible high-performance data visualization for {Julia}},
-  journal = {Journal of Open Source Software}
-}
-
