### 2025-05-27 ebc967c
Update paper.md, Close #85
