### 2025-07-07 #8545 @editorialbot
[REVIEW]: Antique.jl: A Julia package on analytical solutions of quantum mechanical equations
**Submitting author:** @ohno (<a href="http://orcid.org/0009-0001-5222-9726">Shuhei Ohno</a>)
**Repository:** https://github.com/ohno/Antique.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.12.0
