### 2025-07-10 ohno/Antique.jl#90 @goerz
Use `DocumenterCitations`
In the context of https://github.com/openjournals/joss-reviews/issues/8545:

The documentation of `Antique` cites a number of papers. It would probably benefit from using the [`DocumenterCitations`](https://github.com/JuliaDocs/DocumenterCitations.jl) plugin to do this. Full disclosure: I am the maintainer of that package.

This is just a recommendation, certainly not a blocking issue for publication in JOSS ;-)

If you need help setting up `DocumenterCitations`, let me know.
