### 2025-07-10 ohno/Antique.jl#91 @goerz
Release 1.0.0 after JOSS acceptance
As part of https://github.com/openjournals/joss-reviews/issues/8545, you will be asked to tag a new release, after the review concludes. This release will be the one referred to by the paper.

I would encourage you to make that release `v1.0.0`. The Julia community encourages the use of [semantic versioning](https://semver.org). If you follow the ideas laid out in that document, then `v0.x` versions are only for initial development, and the "versioning contract" starts with `v1.0`, which first specifies what constitutes the public API.

[As the FAQ states](https://semver.org/#how-do-i-know-when-to-release-100):

> If your software is being used in production, it should probably already be 1.0.0. If you have a stable API on which users have come to depend, you should be 1.0.0. If you’re worrying a lot about backward compatibility, you should probably already be 1.0.0.

With the formal review of a JOSS publication, which should raise any potential issues with the API, and with a publication being a strong indicator that you want people to use (and cite) the package, you're probably at the stage where you indeed want to provide a _stable_ API, and are in a clear position to do so.

This is obviously not a requirement, just a suggestion.

P.S.: Don't tag a `v1.0.0` release until the JOSS review is concluded (you will be prompted by the editor when it is time)
