### 2025-07-10 ohno/Antique.jl#92 @goerz
Collect and show coverage information
In order to determine whether the tests sufficiently verify the functionality of the package in https://github.com/openjournals/joss-reviews/issues/8545#issuecomment-3045600098, can you make sure that in your [`CI.yml`](https://github.com/ohno/Antique.jl/blob/main/.github/workflows/CI.yml), you collect and upload coverage information to [Codecov](https://about.codecov.io)?

Basically, after the `julia-runtest` step, you should add

```
      - uses: julia-actions/julia-processcoverage@v1
      - uses: codecov/codecov-action@v5
        with:
          files: lcov.info
          token: ${{ secrets.CODECOV_TOKEN }}
```

Make sure you have the `CODECOV_TOKEN` set up as a repository secret, see [their documentation](https://docs.codecov.com/docs/adding-the-codecov-token).

The coverage should be shown as a badge in the README.
