### 2025-07-11 ohno/Antique.jl#90 @lhapp27
Thank you for your comment (and for the package!). I have used DocumenterCitations.jl in the documentation of my own package before, and I agree that it is very convenient for managing references.

@ohno: do we have a complete .bib file of all the references? I think that would be required.
