### 2025-07-12 ohno/Antique.jl#98 @goerz
Use typed fields in models
https://github.com/ohno/Antique.jl/blob/ad463ca173329f7c584ff3315a93950c5dba999f/src/HydrogenAtom.jl#L4-L10

has all fields in the `HydrogenAtom` types as `Any`. Even if you don't care about performance, this is not a good idea:

```julia
julia> HA = HydrogenAtom(Z="1")
Antique.HydrogenAtom(Z=1, mₑ=1.0, a₀=1.0, Eₕ=1.0, ħ=1.0)

julia> E(HA)
ERROR: MethodError: no method matching -(::String)
```

My recommendations would be [the proof-of-concept implementation](https://github.com/goerz-testing/AntiqueModules.jl/blob/ee9cc51d7ef37567a86bb358168b5c59724bc86e/src/models/HydrogenAtoms.jl#L8-L41)

If you want to support both a full ASCII and a Unicode API, then I don't think you can use `@kwdef`. If you stick to ASCII (which I might, at the end of the day, just because it put so much less demand on contributors of new models), and `@kwdef` is probably possible.

This applies to all models, obviously, not just `HydrogenAtom`

Ref. https://github.com/openjournals/joss-reviews/issues/8545#issuecomment-3045600098
