### 2025-07-12 #8545 @goerz
I have concluded my initial review of the package.

First, since [this came up in the pre-review](https://github.com/openjournals/joss-reviews/issues/8237#issuecomment-3001329274), I have no issues whatsoever regarding

> - [x] **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

Models with analytical solutions are quite often used as approximations, even outside of teaching. Having a curated list of models with access to analytical eigenfunctions and other properties could be extremely useful. I would see myself using this package as part of my research.

The manuscript itself seems ready for publication – some minor editing has already been applied in https://github.com/ohno/Antique.jl/pull/89

My main concern, as expressed in https://github.com/ohno/Antique.jl/issues/97, is with

> - [ ] **Functionality:** Have the functional claims of the software been confirmed?

I have some concerns about the design of the API, specifically the use of single-letter function names and the mandatory use of Unicode. Both violate what I believe to be a broad consensus for best practices in the Julia community. Of course, questions of style are never accepted universally, and usually, it is the author's prerogative to determine the design of their package. However, the consensus around these stylistic choices has good reasons behind it, which I've tried to articulate in https://github.com/ohno/Antique.jl/issues/97 and the discussions linked therein. In particular, I think that the current API design limits the potential of the package with respect to the goals stated in the [README](https://github.com/ohno/Antique.jl/blob/main/README.md),

> Antique.jl provides self-contained, well-tested, and well-documented implementations of **an**aly**ti**cal solutions to solvable **qu**antum m**e**chanical models. […] We aim to support researchers, lecturers, students, and any person who is interested in quantum mechanics.

Using descriptive function names as opposed to symbols for function names is an inherent form of self-documentation, as the meaning of the symbols is neither obvious, nor are the particular symbols chosen entirely universal. I would consider it a barrier to entry to the audience targeted by the package. Similarly, the _forced_ use of Unicode can be a usability barrier. Especially students have diverse setups and often a lack of technical skill required to overcome these hurdles. Also, to the extent that these students are then exposed to code not following widely accepted best practices, they may then adopt similar styles in their own code, which is detrimental to the wider ecosystem. A package aimed at teaching should aim to be an example of best practices itself.

I hope that the arguments I'm putting forth here and in https://github.com/ohno/Antique.jl/issues/97 are convincing to the package authors. I understand that I'm asking for a considerable amount of renaming and refactoring, but this review, and the release of a 1.0 version (https://github.com/ohno/Antique.jl/issues/91) seem like the best time to revisit, reorganize, and polish the API of the package. I have created a proof of concept that illustrates my recommendations at https://github.com/goerz-testing/AntiqueModules.jl

For the time being, I have left the following items unchecked:

> - [ ] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I would consider the existing ["API reference" page](https://ohno.github.io/Antique.jl/stable/API/) to be less than useful. As discussed in https://github.com/ohno/Antique.jl/issues/97 and illustrated in the [proof of concept](https://ohno.github.io/Antique.jl/stable/API/), I would recommend separating function and method docstrings, and only show the general function-docstring in the API reference. The docstrings for the model-specific methods are best kept in sections of the documentation specific to that model – as they already are in the existing documentation.

So, besides the minor point of cleaning up the general API page, I'm only keeping this item open to have another look after any potential changes in reaction to https://github.com/ohno/Antique.jl/issues/97 have been made.

> - [ ] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

The tests seem to be in order, but I would ask for the collection of coverage information, see https://github.com/ohno/Antique.jl/issues/92, in order to ensure that the tests cover the entirety of the library. I suspect they already do, but keeping track of coverage and making sure it stays > 80% is a good idea, especially in a dynamic language like Julia, where invalid code only throws errors when it is actually being run in a test.

> - [ ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

There are sufficient community guidelines in place already, but if the recommendations for the restructuring of the library into submodules in https://github.com/ohno/Antique.jl/issues/97 are adopted, these guidelines should be adjusted accordingly.
