### 2025-07-20 ohno/Antique.jl#97 @ohno
@goerz Thank you so much for this incredibly thorough and constructive feedback.

I completely agree with every point about method names and Unicode. This has also been pointed out in other issues:

- https://github.com/ohno/Antique.jl/issues/3#issuecomment-1862962605
- https://github.com/ohno/Antique.jl/issues/12
- https://github.com/ohno/Antique.jl/issues/99

This is exactly the right moment to make these breaking changes as part of our JOSS review cycle. Thank you again for pushing us toward a more user‑friendly API!

The reason why the current code has a "flat" structure is because re-exporting failed in nested modules in the past. I will try to use your proof of concept as a reference.
