### 2025-07-22 ohno/Antique.jl#90 @goerz
> is it possible to display the citation information within the docstring itself?

In principle, yes, as discussed in ["Citations in docstrings"](https://juliadocs.org/DocumenterCitations.jl/stable/syntax/#Citations-in-docstrings).

In your example, since you would be making extensive use of [notes](https://juliadocs.org/DocumenterCitations.jl/stable/syntax/#Syntax-for-Citations) like "p.595 (a.4), (a.6)", you might experiment with using `@citet` or `@citet*`, or, else, using complete custom strings as recommend in ["Citations in docstrings"](https://juliadocs.org/DocumenterCitations.jl/stable/syntax/#Citations-in-docstrings).

I've had some thoughts about adding a `@citefull` feature to DocumenterCitations that would render not just the author name like `@citet`, but a minimal but _complete_ citation. That could be useful here, but I'm probably not going to add this feature in the very near future.
