### 2025-07-25 ohno/Antique.jl#90 @lhapp27
Ok, I have pushed the branch `doc/90` now. It contains the updated make.jl and the refs.bib. The docstrings for DeltaPotential, PoschlTeller and InfinitePotentialWell3D are already updated.

One thing to note: In the PoschlTeller docs I removed one link to "archive . org". I am not certain about the legality of all contents of this website, so I'd rather not distribute their links via our package. If you are certain about it, feel free to put the link back in.
