### 2025-08-08 2412aef
updated references for all models (besides HarmonicOscillaor, and HydrogenAtom)
