### 2025-08-08 ohno/Antique.jl#90 @lhapp27
I updated the references for all models besides HarmonicOscillator and HydrogenAtom, and pushed to the branch `doc/90`. For the SphericalOscillator, one reference was a dead link (beit.tech). I could not recover their blog post, hence I removed the reference.
