### 2025-08-08 3cbab21
add infinite potential well usage
