### 2025-08-08 dc45941
edit hydrogen atom usage
