+++ previews/PR105/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"PoschlTeller/#Pöschl-Teller-Potential","page":"Pöschl-Teller Potential","title":"Pöschl-Teller Potential","text":"","category":"section"},{"location":"PoschlTeller/","page":"Pöschl-Teller Potential","title":"Pöschl-Teller Potential","text":"The Pöschl-Teller potential is one of the few potentials for which the quantum mechanical Schrödinger equation has an analytical solution. It has a finite number of bound states, which can be inferred easily from its potential strength parameter λ. It is defined for one-dimensional systems.","category":"page"},{"location":"PoschlTeller/#Definitions","page":"Pöschl-Teller Potential","title":"Definitions","text":"","category":"section"},{"location":"PoschlTeller/#Antique.PoschlTeller-PoschlTeller","page":"Pöschl-Teller Potential","title":"Antique.PoschlTeller","text":"Model\n\nThis model is described with the time-independent Schrödinger equation\n\n  hatH psi(x) = E psi(x)\n\nand the Hamiltonian\n\n  hatH =\n  - frachbar^22 m fracmathrmd^2mathrmdx ^2\n  - frachbar^2m x_0^2 fraclambda(lambda+1)2 frac1mathrmcosh^2(xx_0)\n\nAfter introducing the dimensionless variables\n\n  x^ast equiv xx_0qquad E^ast equiv frachbar^2m x_0^2 E\n\nthe Schrödinger equation reduces to\n\n  hatH^ast psi(x^ast) = E^ast psi(x^ast)\n\nwith\n\n  hatH^ast = - frac12 fracmathrmd^2mathrmdx^ast^2 - fraclambda(lambda+1)2 frac1mathrmcosh^2(x^ast)\n\nParameters are specified within the following struct:\n\nPT = PoschlTeller(λ=1, m=1.0, ℏ=1.0, x₀=1.0)\n\nlambda determines the potential strength. Currently only integer values for lambda are supported.\n\nReferences\n\nG. Pöschl, E. Teller, Zeitschrift für Physik, 83 (3–4), 143 (1933): More general definitions are given as (2a), (2b) or (11).\nS. Flügge, Practical Quantum Mechanics (Springer Berlin Heidelberg, 1999) p.94 Problem 39. Potential hole of modified Poschl-Teller type.\n\n\n\n\n\n","category":"type"},{"location":"PoschlTeller/#Potential","page":"Pöschl-Teller Potential","title":"Potential","text":"","category":"section"},{"location":"PoschlTeller/#Antique.V-Tuple{PoschlTeller, Any}-PoschlTeller","page":"Pöschl-Teller Potential","title":"Antique.V","text":"V(model::PoschlTeller, x)\n\nbeginaligned\nV(x)\n= -frachbar^2m x_0^2 fraclambda(lambda+1)2 mathrmsech^2(xx_0) \n= -frachbar^2m x_0^2 fraclambda(lambda+1)2 frac1mathrmcosh^2(xx_0) \nendaligned\n\n\n\n\n\n","category":"method"},{"location":"PoschlTeller/#Maximum-Quantum-Number","page":"Pöschl-Teller Potential","title":"Maximum Quantum Number","text":"","category":"section"},{"location":"PoschlTeller/#Antique.nₘₐₓ-Tuple{PoschlTeller}-PoschlTeller","page":"Pöschl-Teller Potential","title":"Antique.nₘₐₓ","text":"nₘₐₓ(model::PoschlTeller)\n\nnote: Note\nNote that the number of bound states nₘₐₓ + 1 is not equal to the maximum quantum number nₘₐₓ, since we count the ground state from n=0 in
[…]
