### 2025-08-08 ef595bd
Update .github/workflows/CI.yml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
