+++ .github/workflows/CI.yml
@@ -33,7 +33,7 @@ jobs:
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
-      - uses: codecov/codecov-action@v4
+      - uses: codecov/codecov-action@v4.6
         with:
           files: lcov.info
           token: ${{ secrets.CODECOV_TOKEN }}
