### 2025-08-08 dea7528
Fix CI.yml for Codecov
