### 2025-08-09 ohno/Antique.jl#90 @ohno
@lhapp27
Thanks for your contribution. I created pull request https://github.com/ohno/Antique.jl/pull/104.
There is a bug with the deployment of the documentation. I will fix it when I commit my local changes.
