+++ docs/src/index.md
@@ -28,16 +28,91 @@ Or run `import Pkg; Pkg.add(; name="Antique", version="0.12.0")` to install on J
 using Antique
 ```
 
-The energy `E()`, the wave function `ψ()`, the potential `V()` and some other functions will be exported. There are two ways to avoid function name conflicts. Run `import Antique` instead of `using Antique`, and use the energy `Antique.E()`, the wave function `Antique.ψ()` and the potential `Antique.V()`. Or try giving other function names like `using Antique: V as potential, E as energy, ψ as wavefuntion, HydrogenAtom`. Here are examples for the hydrogen-like atom. The analytical notation of the energy (the eigen value of the Hamiltonian) is written as
+The energy `E()`, the wave function `ψ()`, the potential `V()` and some other functions will be exported. There are two ways to avoid function name conflicts. Run `import Antique` instead of `using Antique`, and use the energy `Antique.E()`, the wave function `Antique.ψ()` and the potential `Antique.V()`. 
+In the current version, one can access the Greek letters in the section [Greek Letters and Symbols](#greek-letters-and-symbols).
+Or try giving other function names like `using Antique: V as potential, E as energy, ψ as wavefuntion, HydrogenAtom`.
+Here we show two examples: (1) the infinite potential well, and (2) the hydrogen-like atom. There are more examples on each model page.
+
+
+### (1) the infinite potential well
+
+First, we demonstrate one of the simplest models: the infinite potential well. We choose `InfinitePotentialWell` as the model.  Then, we specify the model parameters as follows:
+
+```julia
+IPW = InfinitePotentialWell(L=1.0, m=1.0, ℏ=1.0)
+```
+
+We can access each parameter as follows:
+```julia
+IPW.L
+# output> 1.0
+IPW.m
+# output> 1.0
+IPW.ℏ
+# output> 1.0
+```
+
+The eigenvalues 
+```math
+E_n = \frac{\hbar^2 n^2 \pi^2}{2 m L^2}
+```
+can be computed as follows:
+```julia
+E(IPW, n=1)
+# output> 4.934802200544679
+E(IPW, n=2)
+# output> 19.739208802178716
+```
+
+One of the important features is the wave function 
+```math
+\psi_n(x) = \sqrt{\frac{2}{L}} \sin \frac{n\pi x}{L}
+```
+for different values of `n` and position `x`. We can plot the wave function as follows:
+
+```julia
+using CairoMakie
+
+# settings
+f = Figure()
+ax = Axis(f[1,1], xlabel=L"$x$", ylabel=L"$\psi(x)$")
+
+# plot
+w1 = lines!(ax, 0..1, x -> ψ(IPW, x, n=1))
+w2 = lines!(ax, 0..1, x -> ψ(IPW, x, n=2))
+w3 = lines!(ax, 0..1, x -> ψ(IPW, x, n=3))
+w4 = lines!(ax, 0..1, x -> ψ(IPW, x, n=4))
+w5 = lines!(ax, 0..1, x -> ψ(IPW, x, n=5))
+
+# legend
+axislegend(ax, [w1, w2, w3, w4, w5], [L"n=1", L"n=2", L"n=3", L"n=4", L"n=5"], position=:lb)
+
+f
+```
+![](docs/src/assets/fig/ipw_wavefunction.png)
+
+
+### (2) the hydrogen-like atom
+
+For the hydrogen atom, one need to choose `HydrogenAtom` model and the parameters can be set as 
+```julia
[…]
