+++ docs/src/refs.bib
@@ -134,7 +134,7 @@
   note         = {Accessed: 2025-08-08}
 }
 
-@misc{USCD,
+@misc{UCSD,
   author       = {UCSD},
   title        = {UCSD Physics 130, Quantum Physics},
   howpublished = {\url{https://quantummechanics.ucsd.edu/ph130a/130_notes/node244.html}},
