### 2025-12-29 #8545 @ohno
Hi @lockwo, apologies for my delayed response. I'm sorry that we missed the timeline I previously mentioned. We have fallen behind on the remaining work, and we will need an additional 1–2 months to complete the outstanding items. I'm glad that I'll be able to return to the open-source work with some dedicated time over the year-end and New Year holidays. We expect to restart work shortly; in the meantime, please feel free to set the review status to paused if you think that's the best approach. I'll let you know as soon as we are ready to resume.
