### 2026-03-01 #8545 @lockwo
Hey @ohno just wanted to check in on how things were going
