### 2026-08-20 #8545 @ohno
Apologies for the long interruption. We have resumed work on the project, including in-person discussions during JuliaCon, and brought the remaining review items to completion.

@lockwo
At the time of submission, all three authors shared an affiliation, but we are now at different institutions. Should the affiliations in the paper reflect those at the time of submission or those at the time of publication?

@goerz
We believe the changes requested in https://github.com/ohno/Antique.jl/issues/97 and https://github.com/ohno/Antique.jl/issues/90 have now been addressed. Could you please take a final look at them?

@mdavezac
We believe all points raised in https://github.com/ohno/Antique.jl/issues/99 have now been addressed. Could you please take a final look?
