### 2026-08-20 #8545 @goerz
I commented with a few smaller followup nitpicks, but fundamentally this looks good to me.

I've checked off all points on the checklist, so as far as I'm concerned, this can be published in JOSS together with an upcoming 1.0 release, without further review.
