### 2026-09-06 #8545 @mdavezac
I'm happy to recommend Antique.jl for publication in JOSS.
